Split api and ops utils
