+++ src/api/RealmApi.js
@@ -2,7 +2,7 @@ import util from 'util';
 import { getTenantURL, getCurrentRealmPath } from './utils/ApiUtils.js';
 import { generateAmApi } from './BaseApi.js';
 import storage from '../storage/SessionStorage.js';
-import { printMessage } from './utils/Console.js';
+import { printMessage } from '../ops/utils/Console.js';
 
 const realmsListURLTemplate = '%s/json/global-config/realms/?_queryFilter=true';
 const realmURLTemplate = '%s/json/global-config/realms/%s';
