+++ src/cli/idp/idp-import.js
@@ -8,7 +8,7 @@ import {
   importProvidersFromFiles,
 } from '../../ops/IdpOps.js';
 import storage from '../../storage/SessionStorage.js';
-import { printMessage } from '../../api/utils/Console.js';
+import { printMessage } from '../../ops/utils/Console.js';
 
 const program = new Command('frodo idp import');
