+++ src/cli/journey/journey-describe.js
@@ -13,7 +13,7 @@ import {
   createProgressBar,
   updateProgressBar,
   stopProgressBar,
-} from '../../api/utils/Console.js';
+} from '../../ops/utils/Console.js';
 
 const program = new Command('frodo command sub');
