+++ src/cli/journey/journey-import.js
@@ -6,7 +6,7 @@ import { getTokens } from '../../api/AuthApi.js';
 import { importJourney } from '../../api/TreeApi.js';
 import { importAllJourneys } from '../../ops/JourneyOps.js';
 import storage from '../../storage/SessionStorage.js';
-import { printMessage } from '../../api/utils/Console.js';
+import { printMessage } from '../../ops/utils/Console.js';
 
 const program = new Command('frodo command sub');
