+++ src/cli/journey/journey-prune.js
@@ -7,7 +7,7 @@ import {
   removeOrphanedNodes,
 } from '../../ops/JourneyOps.js';
 import storage from '../../storage/SessionStorage.js';
-import { printMessage } from '../../api/utils/Console.js';
+import { printMessage } from '../../ops/utils/Console.js';
 
 const program = new Command('frodo journey prune');
