+++ src/cli/script/cmd.js
@@ -14,14 +14,14 @@ import {
   convertBase64TextToArray,
   convertTextArrayToBase64,
   validateImport,
-} from '../../api/utils/ExportImportUtils.js';
+} from '../../ops/utils/ExportImportUtils.js';
 import storage from '../../storage/SessionStorage.js';
 import {
   printMessage,
   createProgressBar,
   updateProgressBar,
   stopProgressBar,
-} from '../../api/utils/Console.js';
+} from '../../ops/utils/Console.js';
 
 export default function setup() {
   const script = new Command('script')
