+++ src/cli/theme/cmd.js
@@ -15,14 +15,14 @@ import {
   getTypedFilename,
   saveToFile,
   validateImport,
-} from '../../api/utils/ExportImportUtils.js';
+} from '../../ops/utils/ExportImportUtils.js';
 import storage from '../../storage/SessionStorage.js';
 import {
   printMessage,
   createProgressBar,
   updateProgressBar,
   stopProgressBar,
-} from '../../api/utils/Console.js';
+} from '../../ops/utils/Console.js';
 
 export default function setup() {
   const themeCmd = new Command('theme')
